feat: Add ProductCart screen and functionality
This commit adds the ProductCart screen and functionality to the app. The ProductCart screen allows users to view the products in their cart and proceed to checkout. The ListProducts component has been updated to include a button to add products to the cart. The Product screen has been updated to display the number of items in the cart and allow users to navigate to the ProductCart screen. The types.ts file has been updated to include the IProductCartProps interface.